Making RowMerger a top level object.
